+++ torchmetrics/image/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.image.fid import FID  # noqa: F401
+from torchmetrics.image.inception import IS  # noqa: F401
 from torchmetrics.image.kid import KID  # noqa: F401