+++ torchmetrics/image/fid.py
@@ -142,7 +142,7 @@ class FID(Metric):
     determines if the images should update the statistics of the real distribution or the fake distribution.
 
     .. note:: using this metric with the default feature extractor requires that ``torch-fidelity``
-        is installed. Either install as ``pip install torchmetrics[image-quality]`` or
+        is installed. Either install as ``pip install torchmetrics[image]`` or
         ``pip install torch-fidelity``
 
     .. note:: the ``forward`` method can be used but ``compute_on_step`` is disabled by default (oppesit of
@@ -183,6 +183,8 @@ class FID(Metric):
             If ``feature`` is set to an ``int`` (default settings) and ``torch-fidelity`` is not installed
         ValueError:
             If ``feature`` is set to an ``int`` not in [64, 192, 768, 2048]
+        TypeError:
+            If ``feature`` is not an ``str``, ``int`` or ``torch.nn.Module``
 
     Example:
         >>> import torch
@@ -205,7 +207,7 @@ class FID(Metric):
         compute_on_step: bool = False,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
-        dist_sync_fn: Callable = None
+        dist_sync_fn: Callable[[Tensor], List[Tensor]] = None
     ):
         super().__init__(
             compute_on_step=compute_on_step,
@@ -223,7 +225,7 @@ class FID(Metric):
             if not _TORCH_FIDELITY_AVAILABLE:
                 raise ValueError(
                     'FID metric requires that Torch-fidelity is installed.'
-                    'Either install as `pip install torchmetrics[image-quality]` or `pip install torch-fidelity`'
+                    'Either install as `pip install torchmetrics[image]` or `pip install torch-fidelity`'
                 )
             valid_int_input = [64, 192, 768, 2048]
             if feature not in valid_int_input:
@@ -232,8 +234,10 @@ class FID(Metric):
                 )
 
             self.inception = NoTrainInceptionV3(name='inception-v3-compat', features_list=[str(feature)])
-        else:
+        elif isinstance(feature, torch.nn.Module):
             self.inception = feature
+        else:
+            raise TypeError('Got unknown input to argument `feature`')
 
         self.add_state("real_features", [], dist_reduce_fx=None)
         self.add_state("fake_features", [], dist_reduce_fx=None)